This version only uses commands to work
The options are not well defined, it can search online from new tv episodes, movies. It then adds them to the SQL db, and uses this db to keep track of the completion of new movies. It can search a given folder and copy new movies you haven't watched yet.